Verfiy configuration

+++ chirp_cffi/high_level.py
@@ -114,10 +114,10 @@ typedef struct ch_identity_s {
 } ch_identity_t;
 
 typedef struct {
-    int             REUSE_TIME;
+    float           REUSE_TIME;
     float           TIMEOUT;
-    int             PORT;
-    int             BACKLOG;
+    uint16_t        PORT;
+    uint8_t         BACKLOG;
     char            CLOSE_ON_SIGINT;
     uint32_t        BUFFER_SIZE;
     char            BIND_V6[16];
